requirements.txt
git+https://github.com/dico-api/dico.git@voice
dico-command
dico-interaction
dico-dp
PyNaCl